Fixes to export
